--- vnext/Microsoft.ReactNative/Views/cppwinrt/BorderEffect.idl
@@ -1,41 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// This file was used to generate the files in the cppwinrt subfolders
-//
-// If you modify this file, be sure to run buildcppwinrt.cmd in this folder
-// to generate new headers
-#include "NamespaceRedirect.h"
-
-namespace Microsoft.ReactNative
-{
-    [version(1)]
-    typedef enum CanvasEdgeBehavior
-    {
-        Clamp = (int)0,
-        Wrap = (int)1,
-        Mirror = (int)2
-    } CanvasEdgeBehavior;
-
-    runtimeclass BorderEffect;
-
-
-    [version(1)]
-    [uuid(31602441-15db-5b4a-98dd-ba4247548b40), exclusiveto(BorderEffect)]
-    interface IBorderEffect : IInspectable
-    {
-        [propget] HRESULT ExtendX([out, retval] Microsoft.ReactNative.CanvasEdgeBehavior* value);
-        [propput] HRESULT ExtendX([in] Microsoft.ReactNative.CanvasEdgeBehavior value);
-
-        [propget] HRESULT ExtendY([out, retval] Microsoft.ReactNative.CanvasEdgeBehavior* value);
-        [propput] HRESULT ExtendY([in] Microsoft.ReactNative.CanvasEdgeBehavior value);
-
-        [propget] HRESULT Source([out, retval] Windows.Graphics.Effects.IGraphicsEffectSource** source);
-        [propput] HRESULT Source([in] Windows.Graphics.Effects.IGraphicsEffectSource* source);
-    };
-
-    runtimeclass BorderEffect
-        : [default] IBorderEffect, Windows.Graphics.Effects.IGraphicsEffect
-    {
-    }
-}
